backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pypdf
unstructured
langchain
langchain-google-genai
langchain-community
chromadb
slowapi